refactor before moving base bot code to lib
