+++ ui/serviceWorker/src/serviceWorker.ts
@@ -65,87 +65,87 @@ self.addEventListener('notificationclick', (e: NotificationEvent) => e.waitUntil
 
 // experimental stuff below
 
-self.addEventListener('message', async (e: ExtendableMessageEvent) => {
-  if (e.data && e.data.type !== 'cache') return;
-  if (e.data.value) {
-    const cache = await caches.open('local');
-    hasLocalCache = Promise.resolve(true);
-    await cacheLocalAssets(cache);
-  } else {
-    await caches.delete('local');
-    hasLocalCache = Promise.resolve(false);
-  }
-});
+// self.addEventListener('message', async (e: ExtendableMessageEvent) => {
+//   if (e.data && e.data.type !== 'cache') return;
+//   if (e.data.value) {
+//     const cache = await caches.open('local');
+//     hasLocalCache = Promise.resolve(true);
+//     await cacheLocalAssets(cache);
+//   } else {
+//     await caches.delete('local');
+//     hasLocalCache = Promise.resolve(false);
+//   }
+// });
 
-self.addEventListener('fetch', (e: FetchEvent) => {
-  if (e.request.method !== 'GET') return;
-  const path = new URL(e.request.url).pathname.match(
-    /^\/local(?:[/?#]?.*)?$|^\/assets\/lifat\/bots\/.+$|\/assets\/npm\/zerofish.+$/,
-  )?.[0];
-  if (!path) return;
-  e.respondWith(hasLocalCache.then(haz => (haz ? fetchLocalCache(e, path) : fetch(e.request))));
-});
+// self.addEventListener('fetch', (e: FetchEvent) => {
+//   if (e.request.method !== 'GET') return;
+//   const path = new URL(e.request.url).pathname.match(
+//     /^\/local(?:[/?#]?.*)?$|^\/assets\/lifat\/bots\/.+$|\/assets\/npm\/zerofish.+$/,
+//   )?.[0];
+//   if (!path) return;
+//   e.respondWith(hasLocalCache.then(haz => (haz ? fetchLocalCache(e, path) : fetch(e.request))));
+// });
 
-async function fetchLocalCache(e: FetchEvent, path: string): Promise<Response> {
-  const cache = await caches.open('local');
+// async function fetchLocalCache(e: FetchEvent, path: string): Promise<Response> {
+//   const cache = await caches.open('local');
 
-  try {
-    if (!self.navigator.onLine) throw new Response('offline', { status: 503 });
-    if (path.startsWith('/assets')) {
-      const rsp = await cache.match(e.request);
-      if (rsp) return rsp;
-    }
-    const netRsp = await fetch(e.request);
-    if (netRsp.status >= 300 && netRsp.status < 400) {
-      const redirectUrl = netRsp.headers.get('Location');
-      if (redirectUrl) return await fetch(redirectUrl);
-    }
-    //
-    if (!netRsp.ok) throw netRsp;
+//   try {
+//     if (!self.navigator.onLine) throw new Response('offline', { status: 503 });
+//     if (path.startsWith('/assets')) {
+//       const rsp = await cache.match(e.request);
+//       if (rsp) return rsp;
+//     }
+//     const netRsp = await fetch(e.request);
+//     if (netRsp.status >= 300 && netRsp.status < 400) {
+//       const redirectUrl = netRsp.headers.get('Location');
+//       if (redirectUrl) return await fetch(redirectUrl);
+//     }
+//     //
+//     if (!netRsp.ok) throw netRsp;
 
-    cache.put(e.request, netRsp.clone());
-    cacheLocalAssets(cache);
+//     cache.put(e.request, netRsp.clone());
+//     cacheLocalAssets(cache);
 
-    return netRsp;
-  } catch (err) {
-    console.log('serving cached content', err);
+//     return netRsp;
+//   } catch (err) {
+//     console.log('serving cached content', err);
 
-    return (
-      (await caches.match(e.request)) ??
-      (err instanceof Response ? err : new Response('bad', { status: 500 }))
-    );
-  }
-}
+//     return (
+//       (await caches.match(e.request)) ??
+//       (err instanceof Response ? err : new Response('bad', { status: 500 }))
+//     );
+//   }
+// }
 
-async function cacheLocalAssets(cache: Cache): Promise<void[]> {
-  const promises: Promise<void>[] = [];
-  const assetPaths: string[] = [];
-  const assets: Record<string, string[]> = await fetch('/bots/assets').then(res => res.json());
+// async function cacheLocalAssets(cache: Cache): Promise<void[]> {
+//   const promises: Promise<void>[] = [];
+//   const assetPaths: string[] = [];
+//   const assets: Record<string, string[]> = await fetch('/bots/assets').then(res => res.json());
 
-  console.log('caching assets');
-  for (const [type, list] of Object.entries(assets)) {
-    for (const key of list) {
-      assetPaths.push(
-        ...(type === 'book'
-          ? [`lifat/bots/book/${key}.bin`, `lifat/bots/book/${key}.png`]
-          : [`lifat/bots/${type}/${key}`]),
-      );
-    }
-  }
+//   console.log('caching assets');
+//   for (const [type, list] of Object.entries(assets)) {
+//     for (const key of list) {
+//       assetPaths.push(
+//         ...(type === 'book'
+//           ? [`lifat/bots/book/${key}.bin`, `lifat/bots/book/${key}.png`]
+//           : [`lifat/bots/${type}/${key}`]),
+//       );
+//     }
+//   }
 
-  for (const path of assetPaths) {
-    const assetRequest = new Request(assetUrl(path));
-    const cachedAsset = await cache.match(assetRequest);
+//   for (const path of assetPaths) {
+//     const assetRequest = new Request(assetUrl(path));
+//     const cachedAsset = await cache.match(assetRequest);
 
-    if (cachedAsset) continue;
-    promises.push(
-      fetch(assetRequest).then(assetRsp => {
-        if (assetRsp.ok) {
-          return cache.put(assetRequest, assetRsp.clone());
-        }
-      }),
-    );
-  }
+//     if (cachedAsset) continue;
+//     promises.push(
+//       fetch(assetRequest).then(assetRsp => {
+//         if (assetRsp.ok) {
+//           return cache.put(assetRequest, assetRsp.clone());
+//         }
+//       }),
+//     );
+//   }
 
-  return Promise.all(promises);
-}
+//   return Promise.all(promises);
+// }
